Group/split code into modules/submodules
Some files got too big.
This refactoring groups related files under
common parent modules.
It also moves context functions registration
from workload to the context module.
