+++ src/stats/histogram.rs
